fix "BIT_GYRO" definition error.
"Register 104" bits 2 to 0 are used to reset the analog and digital signal paths of the gyroscope, accelerometer, and temperature sensors. so that 'BIT_GYRO' should be 0x04 instead of 3.
reference: https://www.invensense.com/wp-content/uploads/2015/02/MPU-6000-Register-Map1.pdf @4.27 Register 104 – Signal Path Reset SIGNAL_PATH_RESET

Comment and code behavior do not match.

The code is set to 2000, but the comment is 1000.