Fix CLI vtx command parsing, range checks, and output order in diff Change to use vtxTableConfig PG settings for validation rather than the current runtime values for bands, channels and power levels. Fixes command parsing when restoring a `diff` as the `vtxtable` runtime values won't be available until after the reboot.
Fix "off by one" range check for power level that was preventing use of the highest power level configured in `vtxtable`.

Reverse the order of the `vtxtable` and `vtx` outputs in a `diff`/`dump` so that `vtxtable` settings will be applied first as the `vtx` settings are dependent on them.