Fix a bug that caused uartTotalRxBytesWaiting() to return a grossly incorrect value when DMA was used.  For example, if one byte was waiting, the function returned one less than the size of the buffer.
The reason the bug didn't cause serious problems is that almost all
calls to this function check only whether the returned value is zero or not.
(The only exceptions I found were in "src/main/telemetry/hott.c".  Perhaps
that code was tested on a target that did not use DMA.)  The case of zero
bytes waiting was the only case in which the correct result was returned.