Avoid false accumulation of duplicated rc detection counter
This commit adds another layer of safety for the D8/D16 detection routine.
Now it looks for consecutive 2 frame long "steps" in the rc input, i.e. a
"staircase" pattern.

I belive this is a very unique sign that the RX duplicates the examined
channel in the output, and it's very unlikely to confuse with any other
signal.