Only enable MSP displayPort when displayport_msp_serial is set to a valid UART Previously the logic would initialize and register MSP displayPort anytime `MSP` was selected as the OSD device or when `AUTO` and no other device was found. This occurred even if MSP displayPort was not workable due to a valid UART not being selected.
The logic now won't fallthough and always enable MSP displayPort unless it can work.

Also changed the logic to only disable the OSD feature at runtime instead of permanently in the config if no OSD device was found. This required adding `featureDisableRuntime()`. The previous logic to disable it in the config made configuration difficult as it required several different tabs, saves, and settings in the CLI to properly enable MSP displayPort (or other OSD devices) and each of these steps could disable to the feature. So the user had to keep reenabling the OSD feature at multiple points. With only disabling it at runtime the OSD task is properly prevented ias expected when no OSD device is available.