Adding gyro temperature for MPU6000 & manual only gyro calibration mode
Added gyro temperature readings for MPU6000.
Added gyro temperature debug output (debug_mode = GYRO_TMP) and a SmartPort telemetry sensor (sensorId = 0x0402), currently available for MPU3050 and MPU6000.
Added the manual only gyro calibration mode (option gyro_cal_manual, avialible in CLI).
This mode should help with the gyro temperature drift, allowing to calibrate a gyro when it is at its working temperature, and not during the power-up (as usual).
Tiny Whoops are expected to be the most affected by the gyro temperature drift due to their tight layout and the thermal interference from CPU, ESCs and VTX.
When this mode is activated (gyro_cal_manual = ON), the automatic gyro calibration procedure during power-up is disabled.
The gyro calibration data is saved to EEPROM after the manual gyro calibration (via the stick command or OSD menu).
This calibration data is used on the subsequent boots instead of the actual automatic gyro calibration.
(The gyro_cal_manual mode can also help with launching drones from moving pads (such as boats), when the gyro calibration is complicated.)
To properly calibrate a gyro in this mode, set gyro_cal_manual = ON, save, launch the drone and wait until the gyro is heated to its typical in-flight temperature (check it via the FrSky SmartPort telemetry sensor (it may also be available via FlySky's iBUS and the old FrSky telemetry), or just wait a few minutes after take-off).
Then land, disarm, perform the manual gyro calibration, then (if you are using a level mode) perform the accelerometer calibration (perform the accelerometer trim later, if necessary).
The gyro calibration data is saved in EEPROM and used until the next manual gyro recalibration.
When a gyro is properly calibrated in the gyro_cal_manual mode, a user may only experience a gyro temperature drift during a short period at the beginning of the flight (when the gyro is warming up) and not for the rest of the flight (when the gyro temperature is rather stable).