Fix missing ELRS GPIO configuration.  GPIO pins need to be in INPUT mode before they can be read via IDR.
The reason it's broken on the H7 is that the default gpio mode is
ANALOG, not INPUT like it is on other MCUs, and that the busy pin is
used before communication to the SX1280 and the busy wait function
returns to early because the GPIO IDR bit is always 0 on the H7.  i.e.
not busy, when it is.

It may also be broken on other MCUs prior to this.