Updated tests and seperated calculation from getter for WhDrawn
The getter function should only get the whdrawn so the calculation was
moved to its own function which is called in the batteryUpdate that runs
at 5hz. This will update the value and be true for all calls to the
value in the future. Test were modified but no real test was added.