Lower time required for attitude recovery after crashes
 - Use Kalman gain instead of a static gain for the accelerometer
 Using Kalman gain allows for fast initial convergence after boot or if the gyro is saturated from a crash.
 If the gyro is saturated, the modeled gyro noise is increased, causing the estimate covariance to increase rappidly which in turn mean that new accelerometer measurements are given more wight until the filter has converged again.

- Avoid decreasing accelerometer gain at large errors
The error between the estimated up direction and the upward direction measured by the accelerometer is calculated as a cross product. For angles with an absolute value of more than 90 degrees the magnitude of the error will decrease, slowing down convergence. A check is added for those angles to ensure that the magnitude of the error always increases when the angle increases.

- Remove special cases for accelerometer gain when the vehicle is disarmed
Previously the accelerometer gain was larger when disarmed, and had additional logic to make convergence faster after boot.
The Kalman gain makes this logic unnecessary. Removing the special cases also makes the AHRS behavior more consistent.

- Adjust accelerometer gain dependent on rotation rate magnitude
The rate of rotation can introduce errors in the the direction of the accelerometer vector.
Either through delay from filters or through the centripetal force.
Lowering the acc measurement gain (by increasing the measurement noise) should decrease the impact of these errors.

- Accelerometer measurment outlier rejection based on the Mahalanobis distance
A significant ammount of invalid accelerometer measurements will slipp through even when using the magnitude of the rotational rates and the accelerometer vector to get ridd of the most obvious ones.
The estimate covariance and the base accelerometer meaurement noise is used to ignore measurements that deviates too much from the current estimate.
The outlier rejection is turned off if the quad is disarmed or in gps rescue mode