Update osd_elements.c to report efficiency in Watt-hours / distance
Changes the OSD efficiency element to reports Watt-hours per unit distance. This will provide the pilot with a better representation of actual efficiency when comparing packs of various voltages and should remain more consistent with changing voltage as the pack depletes.

In discussion with long range pilots it is agreed that Watt hours is the correct unit to use: "or example: my 6s 4000mAh pack is 88.2 Wh. Would be nice to see values on screen like 62.38 Wh consumed rather than 6238x mWh"