Add .astylerc config file
Follow the documentations on how the current formatter works.
By adding this config file the formatter being used is more visible, and modifications are directly in the code instead of the docs.

In the docs how ever I would need to adjust the command to be like this:

astyle --options=.astylerc