altitude hold configurable parameter is added for scaling the negative PID output
The PID adjusts the throttle to maintain the target height. Still, it is hard to find the PID parameters that can fix the positive and negative errors with the same parameter set because, in the negative direction, gravity is already helping the system reach the target height, while in the positive direction, it is against it. The added configurable parameter scales down the output of the PID when negative since gravity already plays a role.