altitude hold adaptive hover throttle option
The hover throttle value depends on the battery's voltage; two configurable parameters are added to do this. The user should set the max_battery_level (0.01 v), representing a fully charged battery; at this voltage, the hover-throttle value set on the failsafe menu should work perfectly. The user then should place the battery_drop_scale, which is used to scale the battery drop into throttle units. So if the max voltage is 1660 (16.6 v) and the current value is 1560 (15.6 v). then the difference is 100. This 100 is multiplied by battery_drop_scale % to be added to the throttle hover set in the failsafe menu and used as throttle hover in the altitude hold mode