Properly handle HSI if selected (#14259)
* Properly handle HSI if selected

In some places, PERSISTENT_OBJECT_HSE_VALUE being zero is used to imply that the MCU should be clocked from HSI, but the code doesn't really follow through and results in an invalid clock tree setup.

The proposed changes should fix this.

* Apply suggestions from code review

Co-authored-by: Mark Haslinghuis <mark@numloq.nl>

* Update system_stm32g4xx.c

HSI should now be implicitly selected when SYSTEM_HSE_MHZ is defined as 0 or omitted in the target config.

USE_CLOCK_SOURCE_HSI is not necessary.

* Update src/platform/STM32/startup/system_stm32g4xx.c

Co-authored-by: Petr Ledvina <ledvinap@gmail.com>

* Update src/platform/STM32/startup/system_stm32g4xx.c

Co-authored-by: Mark Haslinghuis <mark@numloq.nl>

---------

Co-authored-by: Mark Haslinghuis <mark@numloq.nl>
Co-authored-by: Petr Ledvina <ledvinap@gmail.com>