Fix compiler errors that appear if you remove -flto=auto from the standard build flags.
Note that, for example, the targets STM32F411 and SITL fail when building with DEBUG=GDB.
This is because of mostly genuine warnings/errors that are hidden by the non-debug
compiler flag -flto=auto (possibly it’s harder for the compiler to spot error situations
when more code is inlined).

A few more errors are revealed when compiling a release build but without -flto=auto.