GYRO: Remove ability for user to set the gyro alignment at run time.
The position of the gyro on the board is NOT user serviceable. The gyro alignment is related to the circuit design, and is the default forward direction of the gyro device relative to the board manufacturers direction indicator (on the board). This cannot be changed by the user.

To prevent undesirable states, this removes the ability for the user to modify the manufacturer settings placed in the respective config.h outside of a custom define (to which we may need to add support for, either in config.h or a new define).

Users can still set the board alignment - which is the mounting orientation of the board manufacturers "direction indicator" (as per silkscreen) relative to the forward direction of the vehicle to which it is installed.

NOTE: The only real reason this was configurable by the user in the past was for "unified targets".

TODO: possibly remove the alignment from PG config and put it in its own array. Not a lot of difference, except a small issue is SENSOR_ALIGNMENT_FROM_STD is not currently "const" - so requires programmatic initialisation of the array.